+++ sdk/keyvault/azure_security_keyvault_keys/perf/get_key.rs
@@ -107,7 +107,7 @@ impl PerfTest for GetKey {
             .unwrap()
             .get_key(self.get_random_key_name(recording), None)
             .await?
-            .into_body()?;
+            .into_model()?;
         Ok(())
     }
 }