+++ sdk/keyvault/azure_security_keyvault_secrets/CHANGELOG.md
@@ -9,6 +9,7 @@
 - `SecretClientListDeletedSecretPropertiesOptions::method_options` is now `PagerOptions`.
 - `SecretClientListSecretPropertiesOptions::method_options` is now `PagerOptions`.
 - `SecretClientListSecretPropertiesVersionsOptions::method_options` is now `PagerOptions`.
+- Renamed `Response<T, F>::into_body(self) -> Result<Response<T>>` to `into_model(self) -> Result<Response<T>>`. `into_body(self)` now returns a `ResponseBody`.
 
 ### Bugs Fixed