+++ sdk/storage/azure_storage_queue/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Breaking Changes
 
 - Changed `QueueServiceClientListQueuesOptions::method_options` from `ClientMethodOptions` to `PagerOptions`
+- Renamed `Response<T, F>::into_body(self) -> Result<Response<T>>` to `into_model(self) -> Result<Response<T>>`. `into_body(self)` now returns a `ResponseBody`.
 
 ### Bugs Fixed